music-analysis/pyproject.toml
[tool.poetry]
name = "mp3-analyzer"
version = "0.1.0"
description = "MP3 Playlist Analyzer & Normalizer with audio feature extraction and playlist management"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "scripts"}]

[tool.poetry.dependencies]
python = "^3.10"
# Audio analysis
librosa = "^0.10.1"
numpy = "^1.26.0"
scipy = "^1.11.0"
# Data manipulation
pandas = "^2.1.0"
# Machine learning
scikit-learn = "^1.3.0"
# Audio download and processing
yt-dlp = "^2023.11.16"
# Playlist APIs
spotipy = "^2.23.0"
ytmusicapi = "^1.3.0"
# Config
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
# Logging
loguru = "^0.7.2"
# Visualization
matplotlib = "^3.8.0"
# Fuzzy matching
rapidfuzz = "^3.5.0"
# Async support
aiohttp = "^3.9.0"
# Data storage
pyarrow = "^14.0.0"
fastparquet = "^2023.10.1"

# 🔤 Collation/sorting dependencies (default: pure Python, no native build)
unidecode = "^1.3.8"
pyuca = "^1.2"

# 🚀 Optional native ICU (install system icu4c first if you want this)
pyicu = { version = "^2.16", optional = true }

[tool.poetry.extras]
# Enable with: `poetry install -E icu` (after installing icu4c on your system)
icu = ["pyicu"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
mypy = "^1.7.0"
black = "^23.11.0"
ruff = "^0.1.6"
types-pyyaml = "^6.0.12"

[tool.poetry.scripts]
# Note: Scripts are organized by pipeline stage in numbered directories
# Run them directly with: poetry run python scripts/XX_stage/script_name.py
preprocess-extract = "scripts.01_preprocess.A_extract_lines:main"
preprocess-clean = "scripts.01_preprocess.B_clean_lines:main"
get-spotify-ids = "scripts.02_create.A_get_ids_for_spotify:main"
get-ytmusic-ids = "scripts.02_create.A_get_ids_for_ytmusic:main"
build-spotify-playlist = "scripts.02_create.B_build_playlists_spotify:main"
build-ytmusic-playlist = "scripts.02_create.B_build_playlists_ytmusic:main"
download-audio = "scripts.03_download.download_audio_yt:main"
analyze-features = "scripts.04_analyze.analyze_features:main"
synthesize = "scripts.05_synthesize.synth_average_track:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
